+++ .atomist/handlers/command/reaction/ReactCommitComment.ts
@@ -66,11 +66,7 @@ class ReactCommitCommentCommand implements HandleCommand {
         const plan = new CommandPlan();
         const execute = { instruction: { kind: "execute", name: "react-github-commit-comment", parameters: this } };
         const msg = "Successfully reacted with :";
-        handleErrors(execute, this);
-        handleSuccess(
-            execute,
-            `${msg}${this.reaction}: to ${this.owner}/${this.repo}/${this.sha1}/comments/${this.comment}`);
-        plan.add(execute);
+        plan.add(handleErrors(execute, this));
         return plan;
     }
 }